working on porting of basic version, removed many things
