+++ DATA_CONFIGS/config_alchemy_full.yml
@@ -1,5 +1,5 @@
 splitter:
-  root: GMDN_SPLITS/
+  root: DATA_SPLITS/
   class_name: pydgn.data.splitter.Splitter
   args:
     n_outer_folds: 1
@@ -7,12 +7,15 @@ splitter:
     seed: 42
     stratify: False
     shuffle: True
-    val_ratio: 0.1
+    inner_val_ratio: 0.1
+    outer_val_ratio: 0.1
     test_ratio: 0.1
 dataset:
   root: DATA/
-  class_name: gmdn_dataset.TUDatasetInterface
+  class_name: pydgn.data.dataset.TUDatasetInterface
   args:
     root: DATA
     name: alchemy_full
     use_node_attr: True
+  transform:
+    - class_name: transform.AlchemyFullPreprocess
