--- GMDN_DATA_CONFIG/config_BARABASI_ALBERT.yml
@@ -1,15 +1,0 @@
-splitter:
-  root: GMDN_SPLITS/
-  class_name: gmdn_splitter.SIRSplitter
-  args:
-    n_outer_folds: 1
-    n_inner_folds: 1
-    seed: 42
-dataset:
-  root: DATA/BARABASI_ALBERT
-  class_name: gmdn_dataset.BarabasiAlbertDataset
-  args:
-    root: DATA/BARABASI_ALBERT
-    name: barabasi_albert
-    size: 100
-    connectivity: 2@5@10@20
