--- GMDN_DATA_CONFIG/config_ERDOS_RENYI.yml
@@ -1,15 +1,0 @@
-splitter:
-  root: GMDN_SPLITS/
-  class_name: gmdn_splitter.SIRSplitter
-  args:
-    n_outer_folds: 1
-    n_inner_folds: 1
-    seed: 42
-dataset:
-  root: DATA/ERDOS_RENYI
-  class_name: gmdn_dataset.ErdosRenyiDataset
-  args:
-    root: DATA/ERDOS_RENYI
-    name: erdos_renyi
-    size: 100
-    connectivity: 0.01@0.05@0.1@0.2
