--- GMDN_MODEL_CONFIG/PROTEINS/DGN/config_DGN_ZINC_full.yml
@@ -1,71 +1,0 @@
-model: gmdn.GMDN
-experiment: pydgn.experiment.supervised_task.SupervisedTask
-log_every: 1
-higher_results_are_better: True
-dataset_getter: pydgn.data.provider.DataProvider
-device: cuda
-num_dataloader_workers: 4
-pin_memory: True
-seed: 42
-grid:
-  supervised_config:
-    batch_size:
-      - 8192
-    shuffle:
-      - True
-    number_of_experts:
-      - 1
-    hidden_units:
-      - 64
-      - 256
-    num_convolutional_layers:
-      - 2
-      - 5
-      - 7
-    dirichlet_alpha:
-      - 1.0
-    epochs:
-      - 1000
-    wrapper:
-      - pydgn.training.engine.TrainingEngine
-    loss:
-      - gmdn_loss.GMDNLoss
-    scorer:
-      - class_name: pydgn.training.callback.score.MultiScore
-        args:
-          # used at model selection time. Should be the one on which to perform early stopping
-          main_scorer: gmdn_score.LikelihoodScore
-          metric2: pydgn.training.callback.score.MeanAverageErrorScore
-          metric3: gmdn_score.DirichletPriorScore    predictor:
-      - gmdn_emission.GraphExpertEmission
-    output_type:
-      - gaussian # binomial # bernoulli # categorical # gaussian # poisson
-    expert_hidden_units:
-      - 0  # not used, linear model
-    neighborhood_aggregation:
-      - add
-    aggregation:  # global aggregation
-      - sum
-      - mean
-      # - null  # for node classification
-    optimizer:
-      - class_name: pydgn.training.callback.optimizer.Optimizer
-        args:
-          optimizer_class_name: torch.optim.Adam
-          lr:
-             0.001
-          weight_decay: 0.  # Useful to constrain magnitude of weights.
-          accumulate_gradients: True
-    early_stopper:
-      #- null
-      - class_name:
-          - pydgn.training.callback.early_stopping.PatienceEarlyStopper
-        args:
-          patience:
-            - 100
-          #monitor: validation_R2 Determination Coefficient
-          monitor: validation_True Log Likelihood
-          mode: max
-          checkpoint: True
-    plotter:
-      - pydgn.training.callback.plotter.Plotter
