+++ MODEL_CONFIGS/config_ZINC_full.yml
